Refatoração Empréstimo Cliente
